Add more function sanitization
